chore: progress를 float로 변경하도록 수정

+++ android/feat/main/src/main/java/com/ssafy/main/encyclopedia/EncycScreenViewModel.kt
@@ -39,7 +39,7 @@ class EncycScreenViewModel
                                                     isDiscovered = it.discovered,
                                                 )
                                             },
-                                        progress = body.data.progress,
+                                        progress = body.data.progress.toFloat(),
                                     )
                                 } ?: EncycScreenState.Error("Failed to load initial data")
                             }
